https://github.com/AndreyAttr/Evo/issues/49 refactoring
